Summary: The is_dynamic flag was added to observers, the relevant
qconfigs and inherited from observers by fake_quants so that it can be
used to determine the quantization type (static vs dyamic) for the
observed module. with the addition of is_dynamic, compute_dtype needs to be refactored out since it does the same job.

Test Plan: python test/test_quantization.py

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D35406334](https://our.internmc.facebook.com/intern/diff/D35406334)

[ghstack-poisoned]

+++ torch/ao/quantization/observer.py
@@ -101,7 +101,7 @@ class ObserverBase(ABC, nn.Module):
     def __init__(self, dtype, is_dynamic=None):
         super(ObserverBase, self).__init__()
         self.dtype = dtype
-        self.is_dynamic=is_dynamic
+        self.is_dynamic = is_dynamic
 
     @abstractmethod
     def forward(self, x):
@@ -1492,7 +1492,7 @@ Per-channel, symmetric weight observer with the 8-bit values restricted to [-127
 """
 
 default_dynamic_quant_observer = PlaceholderObserver.with_args(
-    dtype=torch.float, compute_dtype=torch.quint8, is_dynamic=True
+    dtype=torch.quint8, is_dynamic=True
 )
 """
 Default observer for dynamic quantization.
